pyproject.toml
[project]
name = "ai-safety-radar"
version = "0.1.0"
description = "Autonomous threat intelligence engine for AI security research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "litellm>=1.45",
    "instructor>=1.3",
    "pydantic>=2.8",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "feedparser>=6.0",
    "huggingface-hub>=0.24",
    "datasets>=2.20",
    "streamlit>=1.37",
    "python-dotenv>=1.0",
    "pandas>=2.3.3",
    "redis>=7.1.0",
    "structlog>=25.5.0",
    "plotly>=6.5.0",
    "fakeredis>=2.33.0",
    "pytest-json-report>=1.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "mypy>=1.11",
    "ruff>=0.5",
]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
